Add ShareInventory activities

+++ app/controllers/team_repositories_controller.rb
@@ -11,6 +11,8 @@ class TeamRepositoriesController < ApplicationController
                                          permission_level: create_params[:permission_level])
 
     if team_repository.save
+      log_activity(:share_inventory, team_repository)
+
       render json: { team_repository: team_repository }, status: :ok
     else
       render json: { team_repository: { message: 'not saved!', errors: team_repository.errors } },
@@ -20,9 +22,15 @@ class TeamRepositoriesController < ApplicationController
 
   # DELETE :team_id/repositories/:repository_id/team_repositories/:id
   def destroy
-    team_repository = @repository.team_repositories.find(destory_params[:id])
-    team_repository.destroy
-    render json: { message: 'destroyed!' }, status: :no_content
+    team_repository = @repository.team_repositories.find_by_id(destory_params[:id])
+
+    if team_repository
+      log_activity(:unshare_inventory, team_repository)
+      team_repository.destroy
+      render json: {}, status: :no_content
+    else
+      render json: { message: 'Can\'t find sharing relation for destroy' }, status: :unprocessable_entity
+    end
   end
 
   private
@@ -43,5 +51,16 @@ class TeamRepositoriesController < ApplicationController
 
   def check_sharing_permissions
     render_403 unless can_manage_repository?(@repository)
+  end
+
+  def log_activity(type_of, team_repository)
+    Activities::CreateActivityService
+      .call(activity_type: type_of,
+            owner: current_user,
+            subject: team_repository.repository,
+            team: current_team,
+            message_items: { repository: team_repository.repository.id,
+                             team: team_repository.team.id,
+                             permission_level: team_repository.permission_level })
   end
 end
